Parser rewrite
Still need to debug a few things, but the parser is a lot simpler.

This comes from two things:
  -- a cleaner API
  -- some worked done in the Lexer
  -- drop support for values and annotations found in textual declaration

So this now can only parse the format output by Jan. But it should be easier to add more features if needed.

The lexer does merging of symbols and strings, dot management...

The treatment of numbers and strings should be checked